Ajuste do ellipsis e organização de estilos
